# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for SUT SE 2563",
        "title": "SUT SA Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/nurse": {
            "get": {
                "description": "list nurse entities",
                "produces": [
                    "application/json"
                ],
                "summary": "List nurse entities",
                "operationId": "list-nurse",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.Nurse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "post": {
                "description": "Create nurse",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create nurse",
                "operationId": "create-nurse",
                "parameters": [
                    {
                        "description": "Nurse entity",
                        "name": "nurse",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.Nurse"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Nurse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/nurse/{id}": {
            "get": {
                "description": "get nurse by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a nurse entity by ID",
                "operationId": "get-nurse",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Nurse ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Nurse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "put": {
                "description": "update nurse by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update a nurse entity by ID",
                "operationId": "update-nurse",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Nurse ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Nurse entity",
                        "name": "nurse",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.Nurse"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Nurse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "delete": {
                "description": "get nurse by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a nurse entity by ID",
                "operationId": "delete-nurse",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Nurse ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ent.Department": {
            "type": "object",
            "properties": {
                "department": {
                    "description": "Department holds the value of the \"department\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the DepartmentQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.DepartmentEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.DepartmentEdges": {
            "type": "object",
            "properties": {
                "EdgesOfDepartment2doctorinfo": {
                    "description": "EdgesOfDepartment2doctorinfo holds the value of the EdgesOfDepartment2doctorinfo edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Doctorinfo"
                    }
                },
                "EdgesOfHistorytaking": {
                    "description": "EdgesOfHistorytaking holds the value of the EdgesOfHistorytaking edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Historytaking"
                    }
                }
            }
        },
        "ent.Doctor": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the DoctorQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.DoctorEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
...
        "ent.Financier": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the FinancierQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.FinancierEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                }
            }
        },
        "ent.FinancierEdges": {
            "type": "object",
            "properties": {
                "EdgesOfBills": {
                    "description": "EdgesOfBills holds the value of the EdgesOfBills edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Bill"
                    }
                },
                "EdgesOfUser": {
                    "description": "EdgesOfUser holds the value of the EdgesOfUser edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.User"
                }
            }
        },
...
        "ent.Historytaking": {
            "type": "object",
            "properties": {
                "bp": {
                    "description": "Bp holds the value of the \"bp\" field.",
                    "type": "integer"
                },
                "datetime": {
                    "description": "Datetime holds the value of the \"datetime\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the HistorytakingQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.HistorytakingEdges"
                },
                "hight": {
                    "description": "Hight holds the value of the \"hight\" field.",
                    "type": "number"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "oxygen": {
                    "description": "Oxygen holds the value of the \"oxygen\" field.",
                    "type": "string"
                },
                "pulse": {
                    "description": "Pulse holds the value of the \"pulse\" field.",
                    "type": "integer"
                },
                "respiration": {
                    "description": "Respiration holds the value of the \"respiration\" field.",
                    "type": "integer"
                },
                "symptom": {
                    "description": "Symptom holds the value of the \"symptom\" field.",
                    "type": "string"
                },
                "temp": {
                    "description": "Temp holds the value of the \"temp\" field.",
                    "type": "number"
                },
                "weight": {
                    "description": "Weight holds the value of the \"weight\" field.",
                    "type": "number"
                }
            }
        },
        "ent.HistorytakingEdges": {
            "type": "object",
            "properties": {
                "EdgesOfDepartment": {
                    "description": "EdgesOfDepartment holds the value of the EdgesOfDepartment edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Department"
                },
                "EdgesOfNurse": {
                    "description": "EdgesOfNurse holds the value of the EdgesOfNurse edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Nurse"
                },
                "EdgesOfPatientrecord": {
                    "description": "EdgesOfPatientrecord holds the value of the EdgesOfPatientrecord edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Patientrecord"
                },
                "EdgesOfSymptomseverity": {
                    "description": "EdgesOfSymptomseverity holds the value of the EdgesOfSymptomseverity edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Symptomseverity"
                }
            }
        },
...
        "ent.Medicalrecordstaff": {
            "type": "object",
            "properties": {
                "Name": {
                    "description": "Name holds the value of the \"Name\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the MedicalrecordstaffQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.MedicalrecordstaffEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
...
        "ent.Nurse": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the NurseQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.NurseEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                },
                "nursinglicense": {
                    "description": "Nursinglicense holds the value of the \"nursinglicense\" field.",
                    "type": "string"
                },
                "position": {
                    "description": "Position holds the value of the \"position\" field.",
                    "type": "string"
                }
            }
        },
        "ent.NurseEdges": {
            "type": "object",
            "properties": {
                "EdgesOfHistorytaking": {
                    "description": "EdgesOfHistorytaking holds the value of the EdgesOfHistorytaking edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Historytaking"
                    }
                },
                "EdgesOfUser": {
                    "description": "EdgesOfUser holds the value of the EdgesOfUser edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.User"
                }
            }
        },
...
        "ent.Patientrecord": {
            "type": "object",
            "properties": {
                "Age": {
                    "description": "Age holds the value of the \"Age\" field.",
                    "type": "integer"
                },
                "Allergic": {
                    "description": "Allergic holds the value of the \"Allergic\" field.",
                    "type": "string"
                },
                "Bloodtype": {
                    "description": "Bloodtype holds the value of the \"Bloodtype\" field.",
                    "type": "string"
                },
                "Date": {
                    "description": "Date holds the value of the \"Date\" field.",
                    "type": "string"
                },
                "Disease": {
                    "description": "Disease holds the value of the \"Disease\" field.",
                    "type": "string"
                },
                "Email": {
                    "description": "Email holds the value of the \"Email\" field.",
                    "type": "string"
                },
                "Home": {
                    "description": "Home holds the value of the \"Home\" field.",
                    "type": "string"
                },
                "Idcardnumber": {
                    "description": "Idcardnumber holds the value of the \"Idcardnumber\" field.",
                    "type": "integer"
                },
                "Name": {
                    "description": "Name holds the value of the \"Name\" field.",
                    "type": "string"
                },
                "Phonenumber": {
                    "description": "Phonenumber holds the value of the \"Phonenumber\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PatientrecordQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PatientrecordEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.PatientrecordEdges": {
            "type": "object",
            "properties": {
                "EdgesOfGender": {
                    "description": "EdgesOfGender holds the value of the EdgesOfGender edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Gender"
                },
                "EdgesOfHistorytaking": {
                    "description": "EdgesOfHistorytaking holds the value of the EdgesOfHistorytaking edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Historytaking"
                    }
                },
                "EdgesOfMedicalrecordstaff": {
                    "description": "EdgesOfMedicalrecordstaff holds the value of the EdgesOfMedicalrecordstaff edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Medicalrecordstaff"
                },
                "EdgesOfPatientrecordPatientrights": {
                    "description": "EdgesOfPatientrecordPatientrights holds the value of the EdgesOfPatientrecordPatientrights edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Patientrights"
                    }
                },
                "EdgesOfPrename": {
                    "description": "EdgesOfPrename holds the value of the EdgesOfPrename edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Prename"
                },
                "EdgesOfTreatment": {
                    "description": "EdgesOfTreatment holds the value of the EdgesOfTreatment edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Treatment"
                    }
                }
            }
        },
        "ent.Patientrights": {
            "type": "object",
            "properties": {
                "PermissionDate": {
                    "description": "PermissionDate holds the value of the \"PermissionDate\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PatientrightsQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PatientrightsEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "insurance_id": {
                    "type": "integer"
                },
                "patientrightstype_id": {
                    "type": "integer"
                }
            }
        },
...
        "ent.Registrar": {
            "type": "object",
            "properties": {
                "Name": {
                    "description": "Name holds the value of the \"Name\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the RegistrarQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.RegistrarEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
...
        "ent.Symptomseverity": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the SymptomseverityQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.SymptomseverityEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "symptomseverity": {
                    "description": "Symptomseverity holds the value of the \"symptomseverity\" field.",
                    "type": "string"
                }
            }
        },
        "ent.SymptomseverityEdges": {
            "type": "object",
            "properties": {
                "EdgesOfHistorytaking": {
                    "description": "EdgesOfHistorytaking holds the value of the EdgesOfHistorytaking edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Historytaking"
                    }
                }
            }
        },
...
        "ent.User": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the UserQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.UserEdges"
                },
                "email": {
                    "description": "Email holds the value of the \"email\" field.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "images": {
                    "description": "Images holds the value of the \"images\" field.",
                    "type": "string"
                },
                "password": {
                    "description": "Password holds the value of the \"password\" field.",
                    "type": "string"
                }
            }
        },
        "ent.UserEdges": {
            "type": "object",
            "properties": {
                "EdgesOfDoctor": {
                    "description": "EdgesOfDoctor holds the value of the EdgesOfDoctor edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Doctor"
                },
                "EdgesOfFinancier": {
                    "description": "EdgesOfFinancier holds the value of the EdgesOfFinancier edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Financier"
                },
                "EdgesOfMedicalrecordstaff": {
                    "description": "EdgesOfMedicalrecordstaff holds the value of the EdgesOfMedicalrecordstaff edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Medicalrecordstaff"
                },
                "EdgesOfNurse": {
                    "description": "EdgesOfNurse holds the value of the EdgesOfNurse edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Nurse"
                },
                "EdgesOfUser2registrar": {
                    "description": "EdgesOfUser2registrar holds the value of the EdgesOfUser2registrar edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Registrar"
                },
                "EdgesOfUserPatientrights": {
                    "description": "EdgesOfUserPatientrights holds the value of the EdgesOfUserPatientrights edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Patientrights"
                },
                "EdgesOfUserstatus": {
                    "description": "EdgesOfUserstatus holds the value of the EdgesOfUserstatus edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Userstatus"
                }
            }
        },
        "ent.Userstatus": {
            "type": "object",
            "properties": {
                "Userstatus": {
                    "description": "Userstatus holds the value of the \"Userstatus\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the UserstatusQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.UserstatusEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
...
        "gin.H": {
            "type": "object",
            "additionalProperties": true
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information"
            }
        },
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "authorizationUrl": "",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "authorizationUrl": "",
...
}